pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "testt-49"
version = "0.1.0"
description = "PubMed-based thematic alignment analysis"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.31",
  "pandas>=2.0",
  "numpy>=1.24",
  "tqdm>=4.66",
  "scikit-learn>=1.3",
  "sentence-transformers>=2.6",
  "matplotlib>=3.7",
  "python-dotenv>=1.0",
  "scipy>=1.10",
  "pyarrow>=15",
  "syntok>=1.4.4",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]